+++ app/[eventSlug]/view-session/view-session.tsx
@@ -8,7 +8,7 @@ import { PencilIcon } from "@heroicons/react/24/outline";
 import { CheckCircleIcon, AcademicCapIcon } from "@heroicons/react/24/solid";
 
 import type { Event, Guest, Session, Rsvp } from "@/db/repositories/interfaces";
-import { getEndTimeMinusBreak } from "@/utils/utils";
+import { getEndTimeMinusBreak, TIME_FORMAT } from "@/utils/utils";
 import { UserContext, EventContext } from "../../context";
 import { CurrentUserModal, ConfirmationModal } from "../../modals";
 import { sessionsOverlap } from "../../session_utils";
@@ -257,11 +257,11 @@ export function ViewSession(props: {
           <span>
             {DateTime.fromJSDate(session.startTime ?? new Date())
               .setZone(event.timezone)
-              .toFormat("EEEE h:mm a")}{" "}
+              .toFormat(`EEEE ${TIME_FORMAT}`)}{" "}
             -{" "}
             {getEndTimeMinusBreak(session)
               .setZone(event.timezone)
-              .toFormat("h:mm a")}
+              .toFormat(TIME_FORMAT)}
           </span>
         </div>
         <div className="flex gap-2">